Task B,C - EDA and File Management
B) Defining the data paths and a function to load 10 random images from the training data for plotting a 2x5 subplot with the original label as title.
C) Creating the folder structure using a function for experiment_small_data and experiment_tiny_data with subfolders for test, train, and val, and manually changing the name of the downloaded extracted data to original_data.